Extracted value writing logic of CMDIParserVTDXML into ValueWriter class
